requirements.txt
absl-py==0.13.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
azure-cognitiveservices-anomalydetector==0.3.0
azure-common==1.1.27
backcall==0.2.0
baytune==0.3.12
bleach==3.3.0
botocore==1.20.100
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cycler==0.10.0
dask==2021.3.0
decorator==4.4.2
defusedxml==0.7.1
distributed==2021.3.0
docutils==0.15.2
entrypoints==0.3
featuretools==0.11.0
fsspec==2021.6.1
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.38.1
h5py==2.10.0
HeapDict==1.0.1
idna==2.10
importlib-metadata==4.5.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
iso639==0.1.4
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langdetect==1.0.9
lightfm==1.16
llvmlite==0.34.0
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mlblocks==0.3.4
mlprimitives==0.2.5
mongoengine==0.23.1
msgpack==1.0.2
msrest==0.6.21
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.4.0
numba==0.51.2
numpy==1.16.6
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
orion-ml==0.2.1
packaging==20.9
pandas==0.24.2
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pymongo==3.11.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-louvain==0.13
pyts==0.10.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
s3fs==0.4.2
scikit-image==0.14.5
scikit-learn==0.20.4
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.7.1
six==1.16.0
smart-open==5.1.0
sortedcontainers==2.4.0
statsmodels==0.12.2
tabulate==0.8.9
tblib==1.7.0
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
toolz==0.11.1
tornado==6.1
tqdm==4.49.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
xgboost==0.90
XlsxWriter==1.4.3
zict==2.0.0
zipp==3.4.1
